Create the smallest working document workflow
This commit is the least I can do to get a working workflow where a
user chooses a format (out of a choice of 1), updates the title and
saves the document.

Nothing in this commit is intended to be set in stone - it's likely
that most of it will change in the coming months. It's here to get
something started and we can start on more stories to build out parts
of the interface.

- There's just 1 format to choose from now
- Note that the routes are a mess, and should probably use `resources`,
but I wanted to keep an open mind on how RESTful the application will
be.
- We'll probably want to use better form helpers to create our forms.
- We'll extract creating and updating of models into service objects -
doing it like this is recipe for a mess