Allow users to choose a document type
This implements the 2-step flow of choosing a document type. The
designs for this I found [on the Google Drive for Q1][1], so the copy
might not match the latest thinking. It should be easy enough to change
though.

Things to note:

- We're using 2 YAML files here. The document_types config file is a
slimmer version of what we spiked to configure the forms. I intend for
it to be extended to contain all the knowledge about formats.
- I've pulled out the choosing format into its own controller. This
probably makes sense, because everything will be related to documents
in this application. If we follow a REST-ish method we'll end up with a
huge controller.
- Added the YAML files to `/app/formats` which may be controversial.
Not sure where to put them though - they're so core to the app that
`config` feels wrong.

I'm expecting follow up PRs to:

- Refactor YAML handling. I've resisted the temptation to refactor the
`YAML.load_file` stuff into classes at this time, because I want to
avoid making the decisions about naming / placing stuff now. For what
it's worth, I'm thinking about something like `DocumentType.all` and
`DocumentType.find("press_release")`.
- Start validating the YAML files and testing them against the schemas.
- Add error handling for when a user doesn't select an option and
submits

[1]:
https://drive.google.com/drive/u/1/folders/1lKOFRdHGV00bM33p-Pgf6iYxJB2g
vj0w