Allow users to be redirect to the application that manages a type
This adds a feature to allow users to be redirected to the application
that manages a document type. This will be helpful because we'll be
able to gradually take over responsibility of formats in this
application.

I'm a bit sad about the `Rails.env.test?` in the routes. For unclear
reasons, Capybara will redirect not actually redirect offsite. This
means that when doing a redirect to the URL, we'll see this error in
the tests:

```
ActionController::RoutingError:
  No route matches [GET] "/government/admin/consultations/new"
```

I've considered a few other options: making this a controller test (I'd
like an integration test for the feature), stubbing the controller,
adding an extra page that contains a button to redirect and just test
that, not testing this, and adding the routes in the spec file (doesn't
seem to work).