Add the "Publishing guidance" page
This page contains information on what belongs on GOV.UK.

The rendering is a bit funky (inline File.read/Govspeak), but we don't
know yet how many static content pages will live in this app (or even
if these pages belong in this app). Once we know a little more, we can
create an abstraction layer that is less ugly. One idea could be is to
add a govspeak renderer, so that the app can render pages with govspeak
just like ERB.

The reason were going with Govspeak rather than HTML in these pages is
twofold:

- We might like to allow content publishers on the team access to these
  pages to update the content. In that case they're more familiar with
  govspeak.
- Govspeak makes sure that correct HTML is used. For example, a user
  can't add `<b>` tags (it will always use `<strong>`).