Add MVP for format-specific content editing
This PR adds the minimum to create a system of format-dependent content
editing. It is the first step to allow us to specify the formats in
this app in a YAML.

It works like this:

- In the the document types YAML we define `fields` for a type
- Fields have a `id`, `label` and `type`
- Using these fields, we build a form with the correct inputs. In this
PR that's just the "govspeak" input, but I imagine we'll have inputs
like dates, dropdowns, booleans and more.
- The input of the form is then used in the controller (with an
additional param sanitisation step)
- The extracted input is then persisted into the `contents` JSON
column, which the forms can access

For future PRs:

- Add more input types
- Add more fields for formats
- Use the `contents` in the payload to the publishing-api
(https://trello.com/c/mVJa1YlK)
- Use `DocumentTypeSchema` to refactor the controllers

This majority of this work comes from a spike into the system
@benthorner and I did last week.

https://trello.com/c/PxtaaObC