Fix environment indicators
Currently integration is broken, because we're passing in the
`ENV["ERRBIT_ENVIRONMENT_NAME"]` into the layout component, which loads
a favicon that doesn't exist:

https://sentry.io/govuk/app-content-publisher/issues/614468696/?query=is
:unresolved

This commit instead calculates the environment more safely, so that the
environment will always be one of production, staging, integration, or
development. By default it will be development.