Fix content editing for formats without fields
This fixes a bug that crashes the edit page for documents with a
document type without fields, like news article.

https://sentry.io/govuk/app-content-publisher/issues/618507813

I assumed that having a `fields=` method would make the `ActiveModel`
initialiser use it (just like `attr_accessor` does). It doesn't, but it
does work if there's a `attr_writer`.