Don't use ActiveModel::Model
This changes how we initialise `DocumentTypeSchema` and
`SupertypeSchema`. I initially added `ActiveModel::Model` for
convenience. Since then we've iterated a bit and are having to override
the initializer to get it to do what we want
(https://github.com/alphagov/content-publisher/pull/36). The
abstraction is therefore not really useful anymore.