Import Whitehall JSON into documents
This takes the big wad of JSON that the export:news_documents task
produce and imports them as documents. As we don't store much
information currently this means that most of that data is ignored.

The test really cuts down the JSON to just the bare minimum for the sake
of brevity in the test.