Redirect to document show page after edit
We should take the user to the document show page when they have
finished their create/update task. The tests then require that the show
page render the contents of the document, which we should implement in
the same generic way that the edit page is constructed.

To support a generic show page, the previous 'input_types' partials with
'fields' using the convention that each field has a partial to show its
content (_field.html.erb) and another to edit it (_field_input.html.erb).