Add the component guide to the application
This adds the component guide to the application. This means we can
start developing components that are only in this application.

Because we've disabled Static in this app
(https://github.com/alphagov/content-publisher/blob/fee9dbc6c25f4b9f7190
3daa3cc3284be6599cae/config/application.rb#L30-L33), the normal GOV.UK
styling won't work. This is not a big issue, because the components
preview correctly. Alex will look at tidying up the component guide so
it's better usable in admin apps as well.