Explicitly fetch the details fields from the contents
Previously we've sent `Document#content` directly into the `details` of
the item. We now need to transform some of the values before inserting
them into the `details`.

This has the upside of being less magic - from now on we'll send all of
the fields specified, not just set in the database.