Convert Govspeak to HTML before sending it to publishing-api
This converts fields of the type "govspeak" into HTML before sending it
to the publishing-api.

The system of transformations is meant to be extensible. if we have
more types, they can be added to the
`perform_input_type_specific_transformations` method. If it grows
sufficiently large, it can be moved into a separate class as well.