Fill in field specific content
First, this fixes a bug in this test - the `@schema.fields` code is
missing an `each`, so it's currently not doing anything.

Secondly, this makes sure that we raise an error if we encounter an
unknown field type. This will come in handy if we add a new field type,
because the tests will fails.

Thirdly, this adds a field-specific test to the application. This would
have caught the error mentioned earlier.

I'm imagining this being extracted into some kind of class or helper if
we have more field types.