Check for base_path instead of reserving a path
We are changing how we confirm if a base path can be used by a new piece of
content from making a PUT request to reserve a path to making a GET request to
check if the path is already in use. We'll then reserve the path once the
content is published to draft instead. The rationale behind this is:

- This stops us from needing to worry about "unreserving" a base path in the
  event that it's no longer needed
- We'll have fewer error cases to handle as we'll only need to make one PUT
  content call to the Publishing API upon publish, instead of making both a PUT
path reservation request and a PUT content request

One potential "gotcha" we will need to address in future work is what happens
in cases where a path becomes unavailable between editing the title and
submitting the draft.