Use Document Type YAML to store path prefix
In order to ensure that the fetch request can send cookies, we have to pass
`credentials: "include"`.

This commit adds a `prefix` field to the Document Types YAML so that we can
construct a base_path that won't be in use by other apps. This also means that
we pass a document ID to the generate-path route so that it can look up the
correct path prefix.