Replace supertype thing as a first-class supertype
Previously a static not-sure supertype was defined in the new documents
controller, which required a proxy SupertypeThing class to handle the
generation and processing of options. My original intention was simply
to move this class into /services to highlight its existance, but
instead this PR is an attempt to remove it altogether by adding a 'path'
option to supertypes, behaving similarly to managed_elsewhere for docs.