Append an integer to paths which already exist
If a document exists with the slugified URL we append an integer.

Example:
  if /news/major-news-story already exists
  and the user enters a news article with the title "Major news story!"
  then we would generate a path /news/major-news-story-1

This happens up to a limit of 5, /news/major-news-story-6 will not be generated.
The ajax endpoint will return a 409 (conflict) in this scenario. Unfortunately
that will write an error to the console which breaks our "fail on js errors"
configuration.

An additional test has been added to test the happy path of generating a slug
and displaying a preview to the user.