Cache linkables for 5 mins and only wait for 1 second
We don't want to wait a long time for linkables to return so we set this
to 1 second and we'll render something different to the user.

This sets a rather simplistic 5 minute cache on linkables using a rather
long 10 seconds for the race_condition_ttl. This means that the first
time this entry is accessed when stale it'll try to load it but if any
other cache accesses occurs while this is accessing they'll continue
using the expired one for up to 10 seconds.