Assign showErrorMessage to a variable to avoid function hoisting
strictmode requires any new functions to be defined at the top of the scope as
presumably this make the function hoisting behaviour of JS more obvious (it
would help me at least!).

In this case we want to capture the surrounding scope so assigning it to a
variable is better than moving the function.