Combine associations types into multi_association
A lot of the associations we're going to add will share the same view
code. The only thing that differs between most associations is the
document_type of the linkable we're trying to let the user edit. Rather
than trying to derive this from the id of the association, it seems more
robust to create a separate mdoel to properly represent associations.