Pass schema through to field partials
The separation between the source of the data (the schema) and the use
of that data in a field partial (name, label) making it hard to remember
which value came from where. Passing the schema through to the partial
and explicitly calling it a schema (it's not a field) makes the link
between data and view more visible. In general we should try to be more
consistent in our variable names to differentiate between a schema from
the config file and something that's created to conform to the schema.