Lint Javascript with StandardJS
This adds the StandardJS linter as advised in
https://github.com/alphagov/styleguides/blob/master/js.md#linting

This uses NPM to install the linter rather than Yarn as our CI servers
do have NPM installed and don't have Yarn. Since Yarn plays nicer with
Rails that is probably a better option to take moving forward if we make
use of node-modules in the app, but this seems a reasonable approach for
now.

This sets up a lint and a lint:fix command, the latter can be used to
fix issues automatically.