Remove redundant spec for document_type_schema
This spec was originally written to ensure we were doing a to_a on the
fields param, but it's very much covered by the feature tests - all of
them fail if it's removed. Rather than copy/paste this for the upcoming
'associations' param, it seems better to remove it but keep the coverage.