Refactor our javascript to make it more modular
We should avoid having all our Javascript in one file. A simple
heuristic to split the code is to follow the layout of the views/
directory, such that most of the code in application.js is now found in
documents/edit.js.

We should continue to restrict our Javascript so that it only executes
on the (parts of) pages where it's actually needed. To help iterate this
pattern, we've pulled out a 'restrict' function, which uses the
data-module technique from govuk_publishing_components to scope the JS.