Refactor edit document associations spec
This refactor ensures that the feature itself is tested without testing the
format-specific configuration.

We create a document_type_schema factory so that we can use a test schema
instead of a real-life one in order to specifically test this feature. In order
to be able to create the schema and not add it to the config, we add a create
method to Document Type Schema.