Ensure user chooses supertype / document type
This prevents the user reaching an error page in the case where they
forget to choose a supertype or a document type, and then click the
'Continue' button when creating a document.

The page to choose a document type was previously a POST route, which
made it awkward to re-visit the page if the user doesn't select a
document type and clicks 'Continue'. To support re-visiting the page,
the route is now a GET route and the form_tag for choosing a supertype
now has the explicit route to get to it, turning off the UTF8 character
set param to keep the resulting URL query params clean.