Use translations instead of hard copy
Using the Rails translations framework allows us to centralise the text
in our views, so that it's easier to find and edit. One problem to solve
here is the organisation of this text into locale files. Another problem
is how to cope the fields having different text for each format.

   - formats/en.yml (guidance)
   - layouts/en.yml (layout-specific text)
   - pages/new_document/en.yml (new_document page-specific text)

The organisation into folders is also mirrored in the translation
hierarchy, which makes naming things easy and allows us to add other
types of translation (e.g. for internal components) without the risk of
conflict or ambiguity. There are a few conventios for page translations.

   - The hierarchy has an extra layer to specify the view (e.g. 'edit')
   - Most pages have a 'title' and an 'actions' (buttons/links) section
   - A couple of pages have a 'flashes' section for success/error text

Format-specific text (guidance) was easier to get working using the
Rails translation mechanism. An alternative would be to add it to the
format definition, which would centralise all of the format-specific
configuration, but at the expense of creating two places where copy
exists and requiring additional modelling/lookup work to extract text.