Start exposing the inner workings of the application
This is a first stab at a documentation approach for this app. The
target of these docs are the people working on it - the PM, DM, user
researchers, designers, content designers, and developers.

My dream for documentation is that content publisher will be able to
"explain itself", rather than needing developers to explain it. It's
mainly inspired by the difficulty we have in understanding the current
workings of Whitehall.

This PR adds two things to `/docs`:

- All the formats. This is to answer the question "What formats are
currently available in this application?". Instead of reading the code,
or clicking through the entire app, team members have access to a
single list.
- All the publication states. This is especially important because some
will be rare or transient, so having a central place to show all the
states and translations can make it clearer to the team how the app
works.

I expect this section to grow as the app becomes more complicated.