Set up webpacker for dev VM
This switches the default webpacker port to be 3232 which is a port that
can be reserved in the development VM Procfile.

This also changes the localhost bindings to be 0.0.0.0 so that this is
available outside of the host that has binded this.

Finally this adds a 300ms poll that webpacker will use to check whether
files have changed, this is necessary when using the dev VM as NFS
doesn't fire the system events when files change. Whereas running this
on a host machine Webpack would watch the events and update accordingly.