Model publication states with their descriptions
After having this discussion for document types and supertypes, we
concluded the copy related to an item should be co-located with the
definition of the item, so that they can both change together. This also
has the advantage of better documenting what each state means.

This also solves a hierarchy issue with the previous translations for
publication states: the convention so far has been to place each
translation in a file based on the page it's being displayed on; states
buck this trend, as they're dynamic and displayed on multiple pages.

By making this change, we're treating the definition of domain concepts
(document types, supertypes, publication states) consistently.