Replace hard-coded text with translations in specs
This helps to decouple the tests from the exact text being shown to the
user, such that the tests are robust against content-only changes.

This comes at the expense of making the tests a little harder to read
due to the calls to the translation framework. I've chosen to keep the
lookups inline rather than creating variables, as this would distract
from the more important state being created within each test method.