Test we only send payloads with a single organisation
We know it doesn't make sense to be sending through repeat organisations in the
publishing payload, this explicitly tests for that case.

Moved the keys in the PublishingApiPayload payload_hash to use string keys
rather than symbols. This makes it consistent with the elements we add to the
payload from our Document object, for instance the contents and associations
hashes which are both keyed on strings.