Integrate application with raven-js
Raven-js is a JS module that is used to send errors to sentry.io. To
enable us to do this we expose our SENTRY_DSN and SENTRY_CURRENT_ENV
environment variables to HTML meta tags. We then read those attributes
in as part of our Javascript initialisation and install Sentry.

We also set Raven as a global variable on window so that it can be
available in the console for testing purposes.