Enable sourcemap devtool for production
By default this is set to nosources-source-map which means we don't
serve any source maps in a production environment. This doesn't seem
like something should concern us though since all our source code is on
Github and it allows us to see the source of any errors in sentry (or
can be debugged in the browser)

This is added to the webpack/production.js file since, as far as I can
tell, there is not an option to add to add this to the
config/webpacker.yml file.

This is not needed in development since that defaults to a devtool of
cheap-module-source-map