Fix links to error.css
These were created with a relative path and thus fail if you get an
error or 404 on a page that is not at a root path. E.g. we'd fail to
load the stylesheet on a page like /documents/id as it would try load
the stylesheet at documents/error.css