Make active storage service configurable by env var
This allows us to configure Heroku to use the local settings for this
rather than hook it up to AWS - Heroku does warn us about using this
since the data wouldn't be available if the dyno is restarted or if we
run multiple dynos. But this seems a pretty low risk scenario for PR
previews.