Enforce double quotes and trailing commas
This is brings us closer to the origin GDS style guide and removes the
ambiguity of which quote and list syntax to use, which has been a
persistent pain. Note that the quoting isn't enforced where it would be
semantically invalid (e.g. '"hello"').

https://github.com/alphagov/styleguides/blob/7a9045bf653e6f12d614217b30d70b7798ece5bd/ruby.md

We should look at moving this into the govuk-lint gem after trialing it
here, so that other apps can benefit.