Reorder associations on summary page
This reorders the associations that are displayed on the summary page. For now,
changing their order in the document_types config file is the simplest way to
achieve this. In future when this list of document_types grows we may want to
think about a more sophisticated way of enforcing ordering.