Create model for Images
This creates a db table for images that are associated with documents.

This has an association to ActiveStorage::Blob directly rather than it
being through ActiveStorage::Attachment. The early experiments we made
with ActiveStorage showed that we needed to interact with Blob anyway
and it doesn't seem that we get a huge amount of value with using the
Attachment model.

I debated with myself for a bit as to whether this should be stored with a
document as JSON or as a separate model. The advantage of JSON being
that it'll be easier to track as part of document history, disadvantage
being lack of structured data and other model niceities. After experimenting
with the JSON I eventually concluded that storing it in JSON was perhaps a
premature optimisation and it would be faster to iterate outside of
that.

Once we know more about the needs for history we can re-evaluate this
decision.