Create UploadedImageService and supporting classes
This creates a service for dealing with uploaded images.

It hardcodes in common concerns such as max file size and minimum
dimensions. Max file size is a somewhat abstract number of 20 MB that
isn't based on any research, just an expectation that this is a high
number and people really wouldn't want to be dowloading > 20 MB files to
edit images. The minimum dimensions are based on the current GOV.UK
image dimensions where we restrict/expect all uploaded image files to be
this size.

It performs some validation on a newly uploaded image to check whether
it is suitable and then normalises the image to fix orientation and
remove and exif data. It also uses a basic cropping system to suggest an
initial crop to the user that would have the image in the correct aspect
ratio.

For testing it adds some image files as fixtures which are used as parts
of the tests.

It is intended that this service will be used before a file is added to
ActiveStorage. It does the necessary checks to determine that a file is
suitable for ActiveStorage and gains the information needed to create an
Image record.