Action to create images
This creates the route and controller to allow uploading an image via a
Javascript request.

A presenter class is introduced to serialise the Image model to JSON.
This is done instead of using the `as_json` on the model with the
intention not to specifically couple that method usage to this form of
presentation.