Render Govspeak preview instead of plain Markdown
This commit changes the markdown-editor component to do an async query
to a new protected API that renders the passed Govspeak as HTML.

This requires the component to have knowledge of the API path via a new
data attribute, which is injected deep inside the component hierarchy.
If we were to move this component outside of the app, it ought to have a
default Govspeak rendering service it can use.

Rendering the markdown asynchronously has two error cases: network error
and application error. There's no evidence that an application error is
possible for this API, so we now only handle the network error in JS.

I also considered renaming markdown-* to govspeak-*, but it appears we
refer to 'govspeak' and 'markdown' interchangeably, so left as-is.