Aspect ratio validator for Images
This validator takes an object that responds to width and height methods
and checks that these width and height values conform to the aspect
ratio set up the Image::WIDTH and Image::HEIGHT constants.

As it expects particular methods to exist it isn't the most reusable of
validators. I think however if we discover a need to use it in different
ways then it shouldn't be too painful to adapt with validator options -
but right now it does feel quite coupled to the model that uses it.