Service to update an images crop
This service takes an image model with parameters and then can validate the
input and update the image model.

The validations are in this class rather than on the model itself to
avoid coupling all our validation logic to our ActiveRecord instances.
This approach can be be good to avoid coupling the model with
complicated partial validation scenarios but I'm not sure that we gain a
whole lot from the separation here so am uncertain as to whether this is
the right path or not.