Add pagination for documents index page
This commit adds prev/next links depending on whether prev/next
documents are available, as well as showing the page numbering.
The page limit is hard-coded at 50, which seems to be normal.

Although it's possible to paginate manually, the logic becomes fiddly
around the edge cases, or if the user enters invalid input. Kaminari is
a popular gem that encapsulates this to avoid re-writing it here.