Add validation rules to Image model
This adds the rules that validate the Image model into the model itself
in a standard ActiveRecord manner.

This is a switch from the approach used before where we were validating
user input and then only storing valid data. This has a knock on effect
that for us to support this nicely the user input and output has to
match the model structure.