Update action on DocumentImagesController
This action is used to update an image model. It currently only accepts
crop parameters but could relatively easily be expanded to include other
fields for an image as and when they're needed.

By using the ImageJsonPresenter unchanged this does put us in a weird
situation where the output data doesn't match the input fields. As if we
want to nicely follow the rails validations the input needs to match
that.