Add association between Document and Creator
We add a one-to-many association between a Creator and a Document. A Creator is
a user that has created a document; a user could have many roles in this app so
we make this distinction of a "Creator" role so that we can track who created a
document. A Creator could create many documents, however a document can only
have one creator, hence the decision to create a one-to-many relationship.

A creator is considered optional for a document as there could be instances
where we import a document from another app that doesn't have a creator.  I've
also decided to set `dependent` to `restrict_with_exception` as if there is an
attempt to delete a user, the document they have created shouldn't also be
deleted.