Show raw markdown on the summary page
We think users will find it confusing to see styled markdown content
next to static page content (e.g. titles), so this commit changes the
preview to show the raw markdown using a new 'formatted_text' component.

The component truncates the raw markdown by lines, for which 3 lines
seems to give a reasonable preview (two paragraphs, or title and a
paragraph). If the text is truncated, then we show an ellipsis at the
end of the last non-whitespace line.