Remove unncessary code from Whitehall News Importer
The `lead_organisations` instance variable is only instantiated once when
importing many documents, which meant that it didn't return the correct lead
organisations for each document. We remove this and return the current
edition's lead organisations instead.

We also remove the `to_import` attr_reader as this instance variable is not
used by anything anymore.