Add state class as selector to avoid flaky test
I got caught by this test failing a number of times for me locally which
seems to be related to a timing issue. The hypothesis is that if the
fetch request doesn't return rendered govspeak before capybara looks up
the rendered govspeak then the test will fail.

This can be confirmed by adding a sleep(1) to the render govspeak
controller.

This resolves this issue by adding a class when govspeak is rendered -
this means the capybara finder will keep trying until the class exists.