Maintain sort order when applying a filter
I'm pretty sure - but could be convinced otherwise - that this is
logical behaviour. If you have applied a sort and then adapt your
filters you'd intuitively expect it to maintain that sort. That seems
consistent with tools such as spreadsheets.

It still seems to make sense to not maintain the sort when "Clear
filter" is done as that feels like a reset action and thus feels more
intuitive to clear than maintain.