Implement change note and update type feature
This implements the change note and update type feature.

- Users can now add a change note and select an update type in the edit
screen.
- When a document is first created, it has a "major" update type by
default and "First published." as change note. This can't be changed.
- The update type and change note are sent to the publishing-api, so
change notes for major updates will be displayed on the site. The
update type is no longer part of the `publish` call.
- The change note is displayed on the summary page as well.
- After a document is published, we reset the update type and change
note.
- I've added a test for the creation of a document, because there
wasn't a feature test that tests all the things that need to be set up
when the document is first created.