Scope methods for user facing state
This creates scope methods for each of the current known user facing
states.

The intention is to use these as part of the document index where there
is a Whitehall quirk that is intended to be kept as a feature. This is
that for when a filter is applied to published documents this returns
results all published items whether they need a review or not.