Deal consistently with errors sending to draft
When a document is saved, we attempt to send it to the publishing-api
as draft. This goes wrong sometimes, either because the publishing-api
isn't available (in development), the publishing-api is slow to respond
(when it's busy) or when we're accidentally sending an invalid payload
to the publishing-api (this should not be possible, but it's likely
that we'll make mistakes in this).

This commit adds a "retry" button to the page when the draft publishing
hasn't succeeded. This isn't a particularly elegant solution, but I
can't think of anything better at the moment.

It's intent is to communicate to the user that something went wrong,
and gives an option to fix the problem.

https://trello.com/c/VkVrVGeB