RFC: possible fix for translating buttons
As discussed in the ADR, we're currently not translating clickable
text, primarily to aid the readability of the tests.

Another approach that helps readability is to extract all the `click_on
"Thing"` calls into a helper. The helper can then refer to the
translated string, but the tests still read like English.