Don't associate a document with a creator by default
Initialising a factory without an optional association seems to be a
more consistent approach for FactoryBot. Doing this should mean we have
less cases where the lack of this association breaks features in
integration despite tests passing

For tests that expect or require a user will instead have to provide
that as part of factory initialisation.