Create new action for choosing lead image
The update lead image action was being used in a couple of different ways. One
place it updates attributes but in this case it was just being used to set the
lead image to the one passed in the params.

This change separates those usages to make the controller more obvious.