Filter documents by state on index page
This adds an autocomplete form field for selecting state on the index
page and filters the documents returned accordingly.

This changes the UserFacingState to fallback to a scope of none for an
invalid state, I don't think it's actually necessary for this to raise
an error - we don't have anything using it that considers that.