Add Jasmin for testing our JavaScript
This is based on the approach taken in govuk_publishing_components,
which uses the 'jasmine' gem locked to a 2.x version - some of our
vendored stuff interacts badly with jasmine version 3.x, but it's not
obvious what's causing this and it doesn't seem worth fixing now.

The errors we're getting at the time of writing are below. Grepping for
'script.src' and 'Object.setPro...' gives a match in the vendored assets
for @github and @webcomponents, but it's unclear how we could avoid
requiring these files in the code, so they aren't loaded in test.

  Error during loading
    Message:
        TypeError: null is not an object (evaluating 'script.src')
    Stack:
        No stack trace present.

    Message:
        TypeError: undefined is not a constructor (evaluating 'Object.setPrototypeOf(_CustomElement.prototype, HTMLElement.prototype)')
    Stack:
        No stack trace present.

  1 spec, 0 failures