Upload cropped files using a StringIO object
In this case we want to download an image (in fact an ActiveStorage::Variant
cropped image) and immediately POST it elsewhere (asset-manager). This shouldn't
require us to deal with the filesystem as we don't need to persist the upload
beyond the HTTP request to send it to asset manager.

Currently the GdsApi::AssetManager#create_asset generally takes a Ruby File
object. In fact the only method it uses from File which isn't on the IO
interface is File#path. Because of this we have introduced an AssetManagerFile
class which delegates to a StringIO object other than to add `filename` and a
`content_type` methods. This allows us to send an upload to asset manager
without writing to disk / dealing with temporary files.