Refactor UploadedImageService
I discovered a rather painful issue with our image uploads. We would
take the UploadedFile object and modify the underlying tempfile with
MiniMagick, then when this image was passed to ActiveStorage somehow the
bytes prior to modification was sent. This seems to be to do with
the IO.copy_stream method that ActiveStorage uses but I couldn't quite
pin down what.

This led me to ponder how this scenario could be tested which could be
done by increasing the responsibilities of the UploadedImageService
class to output the eventual model.

While doing this I thought it'd be a good opportunity to rename the
class to be more fitting (since we're having less and less *Service
classes) - it took me ages - I eventually went with ImageUploader as
this uploads an image to a backend storage.

This refactor now has a failing test for stripping the EXIF data which
reflects the problem we've experienced using this with the app.