Change the ImageNormaliser class to output to a new image
Modifying the file that is there has caused some frustrating problems
with ActiveStorage. Simplest approach to resolving this seems to be just
writing the changes to a new Tempfile that can be discarded.

This has the plus that it makes the tests simpler since we don't have to
worry about accidentally modifying a fixture file.