Save Asset Manager file URL to Image
When an image is created we upload the thumbnail variant to Asset Manager and
save the file URL that is returned to the Image record so that we can use it a
later stage.

We only expect to use the Asset Manager Service upload method when creating an
image and so we're happy for it to return the file URL only. We also choose to
set the asset_manager_file_url before saving the Image to the database so that
we can avoid writing to the database twice.