Fix specs not working due to missing stub on index page
The index page now makes a call to the Publishing API to get a list of
organisations to filter by, which means any spec visiting this page must
now support this stub. As there are quite a few of these, it's easier to
add this to a before block that applies to all features.