Handle a failed Asset Manager upload
This handles a failed Asset Manager upload gracefully and returns an error to
the user in this event. We have chosen not save the new image to the database
when this happens, as it would mean that no asset manager file URL is saved to
the image, causing issues when it comes to updating or publishing the image in
future.