Upload to Asset Manager when creating an Image
In a previous commit we introduced uploading a lead image to Asset Manager upon
creation, so we extend this to Images in general. We have also changed the
Image Uploader service to return an instance of an Image but not save it to the
database so that we could avoid having to write to the database a second time
when saving the returned Asset Manager file URL. In this commit we adapt tests
which broke as a result of this change.