Image Uploader Service returns an Image instance
This changes the Image Uploader service to return an instance of an image which
will later be saved. This is so that we can assign `asset_manager_file_url`
after the image has been uploaded to Asset Manager and avoids a situation where
we would need to save to the database twice in fairly quick succession.