Create MVP for history feature
This is a first PR for the "history" feature. It's very rudimentary:

- Whenever a relevant action takes place, we create a `TimelineEntry`
record. I've called this a "timeline entry", because I couldn't come up
with a better generic name. We expect that this page will also include
comments, so "timeline entry" seems appropriate for that too.
- We display the timeline entry in the history tab.

Note that this doesn't implement all of the things from the Trello
ticket, there will be follow up PRs (see "To do" checklist).

https://trello.com/c/TVcThaKi