Handle exception when publishing draft document
This handles exceptions when a draft update is sent to the Publishing API as a
result of a change to a document's lead image.

We have chosen to display an error to users in this case, however it's worth
noting that when this exception occurs our database may become out of sync with
the Publishing API. For example, if a user changed the lead image from
`lead_image1` to `lead_image2` but the update to the publishing-api failed our
database will say the document has `lead_image2` but publishing-api will say
`lead_image1`.